+++ custom_components/solarcharger/models/model_schedule_data.py
@@ -56,8 +56,11 @@ class ScheduleData:
     # Energy management
     max_consumed_energy: float = -1
     consumed_energy: float = 0
-    end_on_max_consumed_energy: bool = False
-    below_max_consumed_energy: bool = False
+    self_paused: int = 0
+
+    # End session if exit condition evaluates to True.
+    end_on_condition: bool = False
+    exit_condition: bool = False
 
     # Requires battery_soc to calculate
     need_charge_duration: timedelta = timedelta.min
@@ -106,9 +109,9 @@ class ScheduleData:
             f"started_max_charge={self.started_max_charge}, max_charge_now={self.max_charge_now}, "
             f"start_next_session_now={self.start_next_session_now}, "
             f"one_percent_charge_duration={self.one_percent_charge_duration}, battery_soc={self.battery_soc}, "
-            f"end_on_max_consumed_energy={self.end_on_max_consumed_energy}, "
-            f"below_max_consumed_energy={self.below_max_consumed_energy}, "
-            f"max_consumed_energy={self.max_consumed_energy}, consumed_energy={self.consumed_energy}, "
+            f"end_on_condition={self.end_on_condition}, "
+            f"exit_condition={self.exit_condition}, "
+            f"max_consumed_energy={self.max_consumed_energy}, consumed_energy={self.consumed_energy}, self_paused={self.self_paused}, "
             f"old_charge_limit={self.old_charge_limit}, new_charge_limit={self.new_charge_limit}, "
             f"calibrate_max_charge_limit={self.calibrate_max_charge_limit}, "
             f"sun_trigger={self.sun_trigger}, "